+++ packages/engine/src/scene/systems/VariantSystem.tsx
@@ -31,8 +31,8 @@ import { defineQuery } from '@etherealengine/ecs/src/QueryFunctions'
 import { defineSystem } from '@etherealengine/ecs/src/SystemFunctions'
 import { SceneState } from '@etherealengine/engine/src/scene/Scene'
 import { EngineState } from '@etherealengine/spatial/src/EngineState'
-import { usePerformanceOffset } from '@etherealengine/spatial/src/renderer/PerformanceState'
 import { MeshComponent } from '@etherealengine/spatial/src/renderer/components/MeshComponent'
+import { usePerformanceOffset } from '@etherealengine/spatial/src/renderer/functions/performanceHooks'
 import { TransformComponent } from '@etherealengine/spatial/src/transform/components/TransformComponent'
 import { Not } from 'bitecs'
 import { useEffect } from 'react'
